msm: ion: Define the source and destination vm parameters as u32
TZ expects the source and destination vm parameters as u32
integers. Define the parameters as u32 instead of u64.

+++ drivers/staging/android/ion/ion_system_secure_heap.c
@@ -59,8 +59,8 @@ static int get_secure_vmid(unsigned long flags)
 static void ion_system_secure_heap_free(struct ion_buffer *buffer)
 {
 	int ret;
-	u64 source_vm;
-	u64 dest_vm;
+	u32 source_vm;
+	u32 dest_vm;
 	struct ion_heap *heap = buffer->heap;
 	struct ion_system_secure_heap *secure_heap = container_of(heap,
 						struct ion_system_secure_heap,
@@ -88,8 +88,8 @@ static int ion_system_secure_heap_allocate(struct ion_heap *heap,
 					unsigned long flags)
 {
 	int ret;
-	u64 source_vm;
-	u64 dest_vm;
+	u32 source_vm;
+	u32 dest_vm;
 	struct ion_system_secure_heap *secure_heap = container_of(heap,
 						struct ion_system_secure_heap,
 						heap);
